requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
anthropic==0.8.1
python-dotenv==1.0.0
